Move XP modules into src
